# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce de productos orgánicos",
        "description": "Documentación para la API de ecommerce de productos orgánicos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend-raices.vercel.app/api/api",
            "description": "Servidor de producción"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener todos los usuarios",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Una lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "id": 1,
                                            "username": "johndoe123",
                                            "email": "johndoe@example.com"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un objeto de usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "username": "johndoe123",
                                    "email": "johndoe@example.com"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado exitosamente"
                    },
                    "403": {
                        "description": "No tienes permiso para eliminar el usuario"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualizar parcialmente un usuario",
                "description": "Permite la actualización parcial de un usuario especificando solo los campos que se desean modificar. Se pueden actualizar todos los campos, excepto email, username y password.",
                "operationId": "6055962f27a3fb88a3445c27693dcf5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "firstname": "Antonio",
                                "lastname": "Banderas",
                                "role_id": 1
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado parcialmente exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "username": "antonio54",
                                    "firstname": "Antonio",
                                    "lastname": "Banderas",
                                    "email": "antoniobanderas@gmail.com"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener el perfil del usuario autenticado",
                "operationId": "c478ac7528255232c445adcd37615e93",
                "responses": {
                    "200": {
                        "description": "Perfil de usuario obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "username": "johndoe123",
                                    "email": "johndoe@example.com"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}